Change async image write API to support scaling, and move functionality to code context.
