+++ praxiscore-video-code/src/main/java/org/praxislive/video/code/VideoCodeDelegate.java
@@ -29,21 +29,15 @@
  */
 package org.praxislive.video.code;
 
-import java.awt.image.BufferedImage;
 import java.util.function.UnaryOperator;
-import javax.imageio.ImageIO;
 import org.praxislive.code.DefaultCodeDelegate;
 import org.praxislive.code.userapi.Async;
 import org.praxislive.core.types.PBytes;
-import org.praxislive.core.types.PError;
 import org.praxislive.video.code.userapi.PFont;
 import org.praxislive.video.code.userapi.PGraphics;
 import org.praxislive.video.code.userapi.PImage;
 import org.praxislive.video.code.userapi.VideoConstants;
-import org.praxislive.video.render.Surface;
 import org.praxislive.video.render.SurfaceOp;
-import org.praxislive.video.render.ops.Blit;
-import org.praxislive.video.render.utils.BufferedImageSurface;
 
 /**
  *
@@ -90,30 +84,32 @@ public class VideoCodeDelegate extends DefaultCodeDelegate {
         context.attachRenderQuery(source, query);
     }
 
+    /**
+     * Write the image as bytes in the specified format. The image will be
+     * encoded asynchronously off the rendering thread.
+     *
+     * @param mimeType mime type of image format
+     * @param image image to write
+     * @return async bytes
+     */
     public final Async<PBytes> write(String mimeType, PImage image) {
-        return write(mimeType, image, 0, 0, image.width, image.height);
+        return context.writeImpl(mimeType, image, image.width, image.height);
     }
 
+    /**
+     * Scale and write the image as bytes in the specified format. The image
+     * will be encoded asynchronously off the rendering thread.
+     *
+     * @param mimeType mime type of image format
+     * @param image image to write
+     * @param width output width
+     * @param height output height
+     * @return async bytes
+     */
     public final Async<PBytes> write(String mimeType, PImage image,
-            double x, double y, double width, double height) {
-        if (!MIME_PNG.equals(mimeType)) {
-            return Async.failed(PError.of(IllegalArgumentException.class, "Unsupported mime type"));
-        }
-        if (image instanceof SurfaceBackedImage surfaceImage) {
-            try {
-                WriteImageSurface wis = new WriteImageSurface(width, height);
-                wis.render(surfaceImage.getSurface(), x, y);
-                return async(wis.getImage(), im -> {
-                    PBytes.OutputStream os = new PBytes.OutputStream();
-                    ImageIO.write(im, "PNG", os);
-                    return os.toBytes();
-                });
-            } catch (Exception ex) {
-                return Async.failed(PError.of(ex));
-            }
-        } else {
-            return Async.failed(PError.of(IllegalArgumentException.class, "Unsupported PImage type"));
-        }
+            double width, double height) {
+        return context.writeImpl(mimeType, image,
+                (int) (width + 0.5), (int) (height + 0.5));
     }
 
     // Start generated PGraphics 
@@ -322,25 +318,4 @@ public class VideoCodeDelegate extends DefaultCodeDelegate {
     }
     // End generated PGraphics
 
-    private static class WriteImageSurface extends BufferedImageSurface {
-
-        private final Blit blit;
-
-        private WriteImageSurface(double width, double height) {
-            super((int) (width + 0.5), (int) (height + 0.5), true);
-            this.blit = new Blit();
-        }
-
-        private void render(Surface source, double x, double y) {
-            blit.setX((int) (x + 0.5));
-            blit.setY((int) (y + 0.5));
-            process(blit, source);
-        }
-
-        @Override
-        protected BufferedImage getImage() {
-            return super.getImage();
-        }
-
-    }
 }
